I made decoration tab more smooth
But to do that I had to give up balance and percentage values changing. Because what I did was just deleting MoveDecorationTab wrapper and adding css animation to the decorationTab div. But now it's impossible to follow it's position inside container and therefore change balance and percentage values when it's needed. But in the technical task there was nothing about that, so I thought that it will be a better solution for now